Fixed bug: gamma was defined as Enzo:ppm:gamma not Enzo:gamma in some input files; debugged checkpoint/restart, which now works with P=1 (FieldDescr::ghosts_, made ItField puppable, set data_ and png_ to 0 when unpacking in OutputImage, made Config puppable, added call to EnzoBlock::initialize in EnzoSimulationCharm pupper)

+++ src/Cello/problem_Problem.cpp
@@ -346,7 +346,7 @@ void Problem::initialize_output
 void Problem::initialize_method(Config * config) throw()
 {
 
-  for (int i=0; i<config->method_sequence.size(); i++) {
+  for (size_t i=0; i<config->method_sequence.size(); i++) {
 
     std::string name = config->method_sequence[i];
